backEnd/services/streamlit-app/requirements.txt
streamlit
requests
pyjwt